Remove polling function and update header function
Remove the polling function since it doesn't really fit with the rest of the lib.

Add retrieveVectorStoreFile function to allowing checking status of individual vector store files.

Update the assistants beta headers function to check if the header has already been added before trying to add it. This avoids issues with the headers growing too large if the header function is called repeatedly, as it would add the same header multiple times